lara/flare: move flare control to TRX
