+++ src/libtrx/config/map.def
@@ -58,6 +58,7 @@ X_CFG_BOOL(gameplay.fix_alligator_ai,                true)
 X_CFG_BOOL(gameplay.fix_bear_ai,                     true)
 X_CFG_BOOL(gameplay.fix_bridge_collision,            true)
 X_CFG_BOOL(gameplay.fix_descending_glitch,           true)
+X_CFG_BOOL(gameplay.fix_flare_throw_priority,        true)
 X_CFG_BOOL(gameplay.fix_floor_data_issues,           true)
 X_CFG_BOOL(gameplay.fix_item_duplication_glitch,     true)
 X_CFG_BOOL(gameplay.fix_lara_pickup_embed,           true)
