+++ src/libtrx/config/map_tr2.def
@@ -1,7 +1,6 @@
 X_CFG_BOOL(audio.enable_lara_mic,               false)
 X_CFG_BOOL(audio.enable_barefoot_sfx,           true)
 X_CFG_BOOL(gameplay.enable_ally_targeting,      true)
-X_CFG_BOOL(gameplay.fix_flare_throw_priority,   true)
 X_CFG_BOOL(gameplay.fix_free_flare_glitch,      true)
 X_CFG_BOOL(gameplay.fix_m16_accuracy,           true)
 X_CFG_BOOL(visuals.fix_glide_cameras,           true)
