+++ src/libtrx/include/libtrx/config/types.h
@@ -267,8 +267,8 @@ typedef struct {
 #else
         bool fix_m16_accuracy;
         bool fix_free_flare_glitch;
-        bool fix_flare_throw_priority;
 #endif
+        bool fix_flare_throw_priority;
         bool fix_walk_run_jump;
         bool fix_wade_wall_hit;
