+++ src/libtrx/include/libtrx/game/lara/flare.h
@@ -3,7 +3,7 @@
 bool Lara_Flare_IsMeshActive(void);
 void Lara_Flare_DrawMeshes(void);
 
-extern void Lara_Flare_Control(void);
-extern void Lara_Flare_Draw(void);
-extern void Lara_Flare_Undraw(void);
-extern void Lara_Flare_Dispose(bool thrown);
+void Lara_Flare_Control(void);
+void Lara_Flare_Draw(void);
+void Lara_Flare_Undraw(void);
+void Lara_Flare_Dispose(bool thrown);
