--- src/tr2/game/lara/flare.h
@@ -1,3 +1,0 @@
-#pragma once
-
-#include <libtrx/game/lara/flare.h>
